Move all constants to single file for ease of maintenance
