--- client/extensions/wp-job-manager/components/setup/constants.js
@@ -1,8 +1,0 @@
-/** @format */
-export const Steps = {
-	CONFIRMATION: 'confirmation',
-	INTRO: 'intro',
-	PAGE_SETUP: 'page-setup',
-};
-
-export const SetupPath = '/extensions/wp-job-manager/setup';
