+++ client/extensions/wp-job-manager/components/setup/setup-redirect/index.jsx
@@ -12,7 +12,7 @@ import page from 'page';
 /**
  * Internal dependencies
  */
-import { SetupPath } from '../constants';
+import { SetupPath } from '../../../constants';
 import { getSiteSlug } from 'state/sites/selectors';
 import { isFetchingSetupStatus, shouldShowSetupWizard } from '../../../state/setup/selectors';
 import QuerySetupStatus from '../../data/query-setup-status';
