+++ client/extensions/wp-job-manager/constants.js
@@ -20,3 +20,11 @@ export const Tabs = {
 		slug: 'pages',
 	},
 };
+
+export const Steps = {
+	CONFIRMATION: 'confirmation',
+	INTRO: 'intro',
+	PAGE_SETUP: 'page-setup',
+};
+
+export const SetupPath = '/extensions/wp-job-manager/setup';
